requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
python-jose
passlib
bcrypt>=4.0.0
jwt
python-multipart
# Database 
pymongo
psycopg2-binary